+++ imgaug/augmenters/pooling.py
@@ -571,12 +571,7 @@ class MedianPooling(_AbstractPoolingBase):
     def _pool_image(self, image, kernel_size_h, kernel_size_w):
         # TODO extend pool to support pad_mode and set it here
         #      to reflection padding
-
-        # we use cval=128 here to decrease the likelihood of unrepresentative
-        # results around the border
-        return ia.pool(
+        return ia.median_pool(
             image,
-            (kernel_size_h, kernel_size_w),
-            np.median,
-            cval=128,
-            preserve_dtype=True)
+            (max(kernel_size_h, 1), max(kernel_size_w, 1))
+        )